Renames parser module
