+++ lib/CSSStyleValue.js
@@ -1,4 +1,4 @@
-import parseAsValue from "./parse-as-value.js";
+import parseAsValue from "./parseCSSText.js";
 
 /**
  * @typedef {object} CSSStyleValue
