+++ lib/parseCSSText.test.js
@@ -1,8 +1,8 @@
 import { describe, it } from "node:test";
 import assert from "node:assert/strict";
-import parseAsValue from "./parse-as-value.js";
+import parseAsValue from "./parseCSSText.js";
 
-describe("#parse-as-value", () => {
+describe("#parseAsValue", () => {
   it("returns CSSUnitValue when expected", () => {
     ["10rem", "20px"].forEach((val) => {
       assert.equal(parseAsValue(val).constructor.name, "CSSUnitValue");
